refactor: 패키지 구조를 계층 중심으로 정리
