+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.config.AppConfig;
+import lotto.common.config.AppConfig;
 import lotto.controller.LottoController;
 
 public class Application {
