+++ src/main/java/lotto/domain/generator/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.generator;
+package lotto.domain.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
