+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import static lotto.exception.ErrorMessage.INVALID_LOTTOS_SIZE;
+import static lotto.domain.exception.ErrorMessage.INVALID_LOTTOS_SIZE;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
